增加将属性转为字符串获取的宏  EXTKeyPathCoding.h、EXTScope.h、metamacros.h 增加 CGWebView 使用 WebKit 库来加载webView 增加UIView+CGCreateConstraint 来创建约束对象 增加具体视图控制器 CGWebViewController

+++ Foundation/NSDate/NSDate+CGTransformString.m
@@ -31,13 +31,13 @@
     NSString *formatterStr = nil;
     switch (stringFormatter) {
         case CGDateForStringFormatterDefault:
-            formatterStr = [NSString stringWithFormat:@"%04li-%02li-%02li %02li:%02li-%02li", (long)dateComponent.year, (long)dateComponent.month, dateComponent.day, (long)dateComponent.hour, dateComponent.minute, dateComponent.second];
+            formatterStr = [NSString stringWithFormat:@"%04li-%02li-%02li %02li:%02li-%02li", (long)dateComponent.year, (long)dateComponent.month, (long)dateComponent.day, (long)dateComponent.hour, (long)dateComponent.minute, (long)dateComponent.second];
             break;
         case CGDateForStringFormatterTimer:
-            formatterStr = [NSString stringWithFormat:@"%02li:%02li-%02li", (long)dateComponent.hour, (long)dateComponent.minute, dateComponent.second];
+            formatterStr = [NSString stringWithFormat:@"%02li:%02li-%02li", (long)dateComponent.hour, (long)dateComponent.minute, (long)dateComponent.second];
             break;
         case CGDateForStringFormatterYear:
-            formatterStr = [NSString stringWithFormat:@"%04li-%02li-%02li", (long)dateComponent.year, (long)dateComponent.month, dateComponent.day];
+            formatterStr = [NSString stringWithFormat:@"%04li-%02li-%02li", (long)dateComponent.year, (long)dateComponent.month, (long)dateComponent.day];
             break;
         default:
             break;
